pyproject.toml
[tool.poetry]
name = "omnipyx-core"
version = "0.1.0"
description = "Omnipyx Core - Modular SaaS Backend"
authors = ["diego1996 <diego.asencio.cuellar@unillanos.edu.co>"]
readme = "README.md"
packages = [{ include = "omnipyx" }]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.2.1"
setuptools = ">=80.9.0,<81.0.0"
djangorestframework = "^3.16.0"
whitenoise = "^6.9.0"
django-cors-headers = "^4.7.0"
django-filter = "^25.1"
gunicorn = "^23.0.0"
psycopg2-binary = "^2.9.10"
django-debug-toolbar = "^5.2.0"
drf-yasg = "^1.21.10"
django-mptt = "^0.17.0"
django-admin-interface = "^0.30.0"
django-colorfield = "^0.14.0"
django-money = "^3.5.4"
drf-spectacular = {extras = ["swagger-ui"], version = "^0.28.0"}
omnipyx-crm = {path = "../omnipyx-crm", develop = true}
django-constance = "^4.3.2"
django-db-multitenant = "^0.3.2"
django-tenants = "^3.8.0"

[build-system]
requires = ["poetry-core>=2.0.0"]
build-backend = "poetry.core.masonry.api"
